mqtt-ingestion/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
paho-mqtt==1.6.1
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
aiosmtplib==3.0.1
twilio==8.11.1
numpy==1.24.3
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
tenacity==8.2.3
